Omogucena obrada predloga i prijem notifikacija
Bibliotekar moze da odobri/odbije predlog (uz obrazlozenje odbijanja), a clan moze da prati status svojih predloga putem in-app notifikacija.
Dodala sam i dve metode, jednu da se poruka oznaci kao procitana, druga da se prebroji koliko korisnik ima neprocitanih poruka.
Ako bude bilo potrebe omogucicu slanej mejla na imejl adresu korisnika.

+++ project/src/main/java/ftn/iis/repository/NotifikacijaRepository.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface NotifikacijaRepository extends JpaRepository<Notifikacija, Long> {
     List<Notifikacija> findAllByKorisnikJmbgOrderByDatumDesc(String jmbg);
-    long countByKorisnikJmbgAndProcitanaFalse(String jmbg);
+    Integer countByKorisnikJmbgAndProcitanaFalse(String jmbg);
 }
